backend/posts/requirements.txt
pyjwt==2.8.0
psycopg2-binary==2.9.9
boto3==1.34.0
